# backend/openapi.json
{
  "components": {
    "schemas": {
...
      "BatchEventRequest": {
        "properties": {
          "event_ids": {
            "items": {
              "type": "integer"
            },
            "maxItems": 500,
            "minItems": 1,
            "title": "Event Ids",
            "type": "array"
          }
        },
        "required": [
          "event_ids"
        ],
        "title": "BatchEventRequest",
        "type": "object"
      },
...
      "DriftHistoryOut": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/DriftSnapshotOut"
            },
            "title": "Items",
            "type": "array"
          }
        },
        "required": [
          "items"
        ],
        "title": "DriftHistoryOut",
        "type": "object"
      },
      "DriftReport": {
        "description": "Envelope so the UI can distinguish 'no drift data' from an actual result.",
        "properties": {
          "available": {
            "title": "Available",
            "type": "boolean"
          },
          "model_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model Name"
          },
          "model_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model Version"
          },
          "reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          },
          "snapshot": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/DriftSnapshotOut"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "available"
        ],
        "title": "DriftReport",
        "type": "object"
      },
      "DriftRunRequest": {
        "properties": {
          "window_hours": {
            "default": 24,
            "maximum": 720.0,
            "minimum": 1.0,
            "title": "Window Hours",
            "type": "integer"
          }
        },
        "title": "DriftRunRequest",
        "type": "object"
      },
...
      "FlowRecordIn": {
        "description": "Single-flow ingest payload used by ``POST /api/v1/ingest/flow``.",
        "properties": {
          "dst_ip": {
            "title": "Dst Ip",
            "type": "string"
          },
          "dst_port": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dst Port"
          },
          "event_time": {
            "format": "date-time",
            "title": "Event Time",
            "type": "string"
          },
          "features": {
            "additionalProperties": true,
            "title": "Features",
            "type": "object"
          },
          "label": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Label"
          },
          "protocol": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Protocol"
          },
          "src_ip": {
            "title": "Src Ip",
            "type": "string"
          },
          "src_port": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Src Port"
          }
        },
        "required": [
          "event_time",
          "src_ip",
          "dst_ip"
        ],
        "title": "FlowRecordIn",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ModelInfoOut": {
        "properties": {
          "algorithm": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Algorithm"
          },
          "artifact_dir": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Artifact Dir"
          },
          "benign_label": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Benign Label"
          },
          "classes": {
            "items": {
              "type": "string"
            },
            "title": "Classes",
            "type": "array"
          },
          "db_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Db Id"
          },
          "feature_order": {
            "items": {
              "type": "string"
            },
            "title": "Feature Order",
            "type": "array"
          },
          "is_active": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Is Active"
          },
          "loaded": {
            "title": "Loaded",
            "type": "boolean"
          },
          "loaded_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Loaded At"
          },
          "metrics_summary": {
            "additionalProperties": true,
            "title": "Metrics Summary",
            "type": "object"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "threshold": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Threshold"
          },
          "version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Version"
          }
        },
        "required": [
          "loaded"
        ],
        "title": "ModelInfoOut",
        "type": "object"
      },
...
      "PredictRequest": {
        "properties": {
          "flows": {
            "items": {
              "$ref": "#/components/schemas/FlowRecordIn"
            },
            "maxItems": 500,
            "minItems": 1,
            "title": "Flows",
            "type": "array"
          }
        },
        "required": [
          "flows"
        ],
        "title": "PredictRequest",
        "type": "object"
      },
      "PredictionOut": {
        "properties": {
          "alert_created": {
            "title": "Alert Created",
            "type": "boolean"
          },
          "alert_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Alert Id"
          },
          "benign": {
            "title": "Benign",
            "type": "boolean"
          },
          "class_probabilities": {
            "additionalProperties": {
              "type": "number"
            },
            "title": "Class Probabilities",
            "type": "object"
          },
          "confidence": {
            "title": "Confidence",
            "type": "number"
          },
          "event_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Event Id"
          },
          "predicted_label": {
            "title": "Predicted Label",
            "type": "string"
          },
          "threshold": {
            "title": "Threshold",
            "type": "number"
          }
        },
        "required": [
          "predicted_label",
          "confidence",
          "class_probabilities",
          "threshold",
          "benign",
          "alert_created"
        ],
        "title": "PredictionOut",
        "type": "object"
      },
...
      "RunRequest": {
        "properties": {
          "limit": {
            "default": 100,
            "maximum": 10000.0,
            "minimum": 1.0,
            "title": "Limit",
            "type": "integer"
          }
        },
        "title": "RunRequest",
        "type": "object"
      },
      "RunSummary": {
        "properties": {
          "alerts_created": {
            "title": "Alerts Created",
            "type": "integer"
          },
          "benign_count": {
            "title": "Benign Count",
            "type": "integer"
          },
          "by_label": {
            "additionalProperties": {
              "type": "integer"
            },
            "title": "By Label",
            "type": "object"
          },
          "model_name": {
            "title": "Model Name",
            "type": "string"
          },
          "model_version": {
            "title": "Model Version",
            "type": "string"
          },
          "processed": {
            "title": "Processed",
            "type": "integer"
          }
        },
        "required": [
          "processed",
          "alerts_created",
          "benign_count",
          "by_label",
          "model_name",
          "model_version"
        ],
        "title": "RunSummary",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "AI-driven intrusion detection and response dashboard.",
    "title": "SentinelAI",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/detection/batch": {
      "post": {
        "operationId": "detect_many_api_v1_detection_batch_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "in": "cookie",
            "name": "sentinelai_access",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Sentinelai Access"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BatchEventRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/PredictionOut"
                  },
                  "title": "Response Detect Many Api V1 Detection Batch Post",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Detect Many",
        "tags": [
          "detection"
        ]
      }
    },
    "/api/v1/detection/drift/history": {
      "get": {
        "operationId": "drift_history_api_v1_detection_drift_history_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "in": "cookie",
            "name": "sentinelai_access",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Sentinelai Access"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DriftHistoryOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Drift History",
        "tags": [
          "detection"
        ]
      }
    },
    "/api/v1/detection/drift/latest": {
      "get": {
        "operationId": "drift_latest_api_v1_detection_drift_latest_get",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "in": "cookie",
            "name": "sentinelai_access",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Sentinelai Access"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DriftReport"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Drift Latest",
        "tags": [
          "detection"
        ]
      }
    },
    "/api/v1/detection/drift/run": {
      "post": {
        "operationId": "drift_run_api_v1_detection_drift_run_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "in": "cookie",
            "name": "sentinelai_access",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Sentinelai Access"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/DriftRunRequest"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Request"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DriftReport"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Drift Run",
        "tags": [
          "detection"
        ]
      }
    },
    "/api/v1/detection/events/{event_id}": {
      "post": {
        "operationId": "detect_one_api_v1_detection_events__event_id__post",
        "parameters": [
          {
            "in": "path",
            "name": "event_id",
            "required": true,
            "schema": {
              "title": "Event Id",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "in": "cookie",
            "name": "sentinelai_access",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Sentinelai Access"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PredictionOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Detect One",
        "tags": [
          "detection"
        ]
      }
    },
    "/api/v1/detection/model": {
      "get": {
        "operationId": "model_info_api_v1_detection_model_get",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "in": "cookie",
            "name": "sentinelai_access",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Sentinelai Access"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelInfoOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Model Info",
        "tags": [
          "detection"
        ]
      }
    },
    "/api/v1/detection/predict": {
      "post": {
        "operationId": "predict_api_v1_detection_predict_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "in": "cookie",
            "name": "sentinelai_access",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Sentinelai Access"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PredictRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/PredictionOut"
                  },
                  "title": "Response Predict Api V1 Detection Predict Post",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Predict",
        "tags": [
          "detection"
        ]
      }
    },
    "/api/v1/detection/run": {
      "post": {
        "operationId": "run_recent_api_v1_detection_run_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "in": "cookie",
            "name": "sentinelai_access",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Sentinelai Access"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RunRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunSummary"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Run Recent",
        "tags": [
          "detection"
        ]
      }
    },
...
  }
}